merge from master

+++ src/Avalonia.Input/Raw/RawMouseWheelEventArgs.cs
@@ -1,8 +1,6 @@
 // Copyright (c) The Avalonia Project. All rights reserved.
 // Licensed under the MIT license. See licence.md file in the project root for full license information.
 
-using System;
-using Avalonia.Layout;
 
 namespace Avalonia.Input.Raw
 {
@@ -10,7 +8,7 @@ namespace Avalonia.Input.Raw
     {
         public RawMouseWheelEventArgs(
             IInputDevice device,
-            uint timestamp,
+            ulong timestamp,
             IInputRoot root,
             Point position,
             Vector delta, InputModifiers inputModifiers)
